Crate twizzler_security

Modules

sec_ctx

Structs

Cap

A capability that represents authorization for a Security Context to access an object.

CapFlags

Del

Gates

Revoc

Specifies when a Capability is invalid. Currenty is a time in ns from unix epoch but plan to change later.

Signature

SigningKey

The Objects signing key stored internally in the kernel used during the signing of capabilities.

VerifyingKey

Enums

DelInner

GatesError

HashingAlgo

SigningScheme