twizzler_security

Struct SecCtxBase

pub struct SecCtxBase {
    pub map: FnvIndexMap<ObjID, Vec<CtxMapItem, MAP_ITEMS_PER_OBJ>, SEC_CTX_MAP_LEN>,
    pub masks: FnvIndexMap<ObjID, Mask, MASKS_MAX>,
    pub global_mask: Protections,
    pub offset: usize,
    pub flags: SecCtxFlags,
}

The base of a Security Context, holding a map to the capabilities and delegations stored inside, masks on targets

Fields
map: FnvIndexMap<ObjID, Vec<CtxMapItem, MAP_ITEMS_PER_OBJ>, SEC_CTX_MAP_LEN>

A map that holds the mapping from an ObjID to the capabilities and delegations stored inside this context that apply to that target ObjID.

masks: FnvIndexMap<ObjID, Mask, MASKS_MAX>

A map holding masks that apply for a target ObjID.

global_mask: Protections

The global mask that applies to all protections gratned by this Security Context.

offset: usize

The running offset into the object where a new entry can be inserted.

flags: SecCtxFlags

Flags specific to this security context.

Implementations

impl SecCtxBase

pub fn new(global_mask: Protections, flags: SecCtxFlags) -> Self

Creates a new SecCtxBase.